refactor(adr-0001): migrate runner out of Next.js, into Dockerfile CMD
Replaces the instrumentation-time auto-migrate with a standalone Node
script wrapped by the production Dockerfile's CMD. Removes the Turbopack
workarounds (@ts-expect-error on .ts ext, duplicated postgres/drizzle-orm
deps in platform-ui) that the in-process approach forced.

- packages/platform-ui/scripts/migrate-postgres.mjs — standalone ESM
  runner. Exits 0 when STORAGE_BACKEND != postgres; otherwise applies
  pending Drizzle migrations via drizzle-orm/postgres-js/migrator.
- packages/platform-ui/Dockerfile — copies migrations folder + script
  to runner stage; CMD = `migrate && server.js`.
- packages/platform-ui/src/instrumentation.ts — reverted to env
  validation only.
- packages/platform-ui/src/instrumentation-postgres.ts — deleted.
- packages/platform-infra/src/postgres/migrator.ts — deleted (dead).
- packages/platform-infra/src/index.ts — drops applyPostgresMigrations
  export.
- package.json — root `pnpm db:migrate` + `pnpm db:generate` scripts
  for local dev.
- README.md / .env.example / docs/postgres-local-dev.md / ADR-0001 /
  PLAN-0001 — local dev = manual `pnpm db:migrate`; Docker prod auto
  via CMD wrapper. Rationale documented as Implementation pattern 6.

Smoke: `pnpm db:migrate` (idempotent), `node migrate-postgres.mjs` with
and without STORAGE_BACKEND. typecheck clean, platform-infra 103/103.
