+++ .env.example
@@ -27,8 +27,10 @@ DOCKER_OPENROUTER_API_KEY=
 # DATABASE_POOL_MAX caps the shared per-process pool. One pool is reused
 # across every PG-backed repository to stay under Postgres' default
 # max_connections=100. Default 10.
-# When STORAGE_BACKEND=postgres, platform-ui auto-applies pending drizzle
-# migrations at boot (idempotent — drizzle's __drizzle_migrations ledger).
+# Migrations: production container CMD wraps migrate + server.js (see
+# packages/platform-ui/Dockerfile + scripts/migrate-postgres.mjs). Local
+# dev runs them manually: `pnpm db:migrate`. Idempotent — drizzle's
+# __drizzle_migrations ledger skips already-applied files.
 DATABASE_URL=postgresql://mediforce:mediforce@localhost:5432/mediforce
 DATABASE_POOL_MAX=10
 STORAGE_BACKEND=firestore
