+++ docs/adr/PLAN-0001.md
@@ -695,41 +695,52 @@ Both files are committed together. The journal is the source of truth for
 "what should be applied"; see the rename rule in §10.1 for branch
 collisions.
 
-**Applying migrations at runtime.** `platform-ui`'s
-`instrumentation.register()` boot hook calls `applyPostgresMigrations(db)`
-([source](../../packages/platform-infra/src/postgres/migrator.ts)) when
-`STORAGE_BACKEND=postgres`. Drizzle's `migrate()` reads the journal, runs
-each pending file inside a transaction, and records it in
-`drizzle.__drizzle_migrations` (schema `drizzle`, table
+**Applying migrations.** Migrations live *outside* the Next.js process —
+they run as a separate Node script in production and as a developer
+command locally. Both paths converge on Drizzle's `migrate()` which reads
+the journal, runs each pending file inside a transaction, and records it
+in `drizzle.__drizzle_migrations` (schema `drizzle`, table
 `__drizzle_migrations` — two underscores, drizzle's own convention).
-Subsequent boots are no-ops for already-applied migrations, so the call
-is safe on every server start.
+Running twice is a no-op for already-applied migrations.
 
-**Why app-boot is safe today.** Production is single-container per
-`docker-compose.prod.yml` (one `platform-ui`, one `container-worker`).
-The `platform-ui` container is the only process that runs migrations.
-There is no race because there is no peer.
+- **Production (Docker):**
+  [`packages/platform-ui/Dockerfile`](../../packages/platform-ui/Dockerfile)
+  wraps the container `CMD`:
+  ```sh
+  node packages/platform-ui/scripts/migrate-postgres.mjs && node packages/platform-ui/server.js
+  ```
+  The script (
+  [`migrate-postgres.mjs`](../../packages/platform-ui/scripts/migrate-postgres.mjs)
+  ) exits 0 when `STORAGE_BACKEND != postgres`, fails the container start
+  on any migration error (fail-stop — half-migrated schema is worse than
+  "container failed to start"), and otherwise applies pending migrations
+  via drizzle's runtime migrator.
+
+- **Local dev:** explicit `pnpm db:migrate` (forwards to drizzle-kit CLI
+  inside `packages/platform-infra`). The dev server does NOT auto-migrate
+  — schema is a deliberate operator step, not hidden in `next dev`. See
+  [`docs/postgres-local-dev.md`](../postgres-local-dev.md).
+
+- **Operator path (pharma on-prem):** same `pnpm db:migrate` invocation
+  if customers want to gate migrations through their change-management
+  process. Or, run the same `migrate-postgres.mjs` script inside their
+  Docker image at any point — the ledger guarantees idempotency.
+
+**Why the script-before-server pattern instead of in-process auto-migrate.**
+An earlier attempt routed migrations through Next.js's
+`instrumentation.register()` boot hook, but Turbopack's instrumentation
+pipeline does not honour `transpilePackages` for workspace imports.
+That forced `@ts-expect-error` escapes on relative `.ts` imports and
+duplicated `postgres` / `drizzle-orm` as `platform-ui` direct deps. The
+standalone-script approach drops both workarounds and matches the
+canonical Rails/Django/Prisma pattern (migrate as a separate process
+step, not a side-effect of app boot).
 
 **Multi-replica future.** When we scale `platform-ui` past one replica,
-two replicas booting simultaneously would race for the migration locks
-drizzle takes internally — usually safe in practice (the second loses,
-sees no pending work, moves on) but not formally guaranteed. The fix is
-a Postgres advisory lock around the `migrate()` call
-(`pg_advisory_lock(<hash('mediforce-migrations')>)` → migrate → unlock).
-Tracked as a follow-up; deferred until we actually run multi-replica.
-
-**Operator path (pharma on-prem).** Customers who want migrations gated
-through their own change-management process can disable auto-migrate
-(they can leave `STORAGE_BACKEND=postgres` and either pre-apply or
-override the boot hook in a downstream build — TBD as we package on-prem
-deliveries) and run migrations by hand:
-
-```bash
-pnpm --filter @mediforce/platform-infra exec drizzle-kit migrate
-```
-
-The drizzle ledger guarantees idempotency, so running this twice or after
-the app has already applied the migration is a no-op.
+the same script becomes a one-shot init-container in compose
+(`depends_on: { migrate: { condition: service_completed_successfully } }`)
+so exactly one process ever runs migrations per deploy. Tracked as a
+follow-up; deferred until we actually run multi-replica.
 
 ---
